repositories modified

+++ src/Garcia.Persistence.EntityFramework/EntityFrameworkRepository.cs
@@ -20,12 +20,14 @@ namespace Garcia.Persistence.EntityFramework
 
         public override async Task<T> GetByIdAsync(long id)
         {
-            return await _dbContext.Set<T>().FindAsync(id);
+            return await _dbContext.Set<T>()
+                .FirstOrDefaultAsync(x => !x.Deleted && x.Id == id);
         }
 
         public override async Task<IReadOnlyList<T>> GetAllAsync()
         {
-            return await _dbContext.Set<T>().ToListAsync();
+            return await _dbContext.Set<T>().Where(x => !x.Deleted)
+                .ToListAsync();
         }
 
         public override async Task<long> AddAsync(T entity)
@@ -40,15 +42,23 @@ namespace Garcia.Persistence.EntityFramework
             return await _dbContext.SaveChangesAsync();
         }
 
-        public override async Task<long> DeleteAsync(T entity)
+        public override async Task<long> DeleteAsync(T entity, bool hardDelete = false)
         {
+            if(!hardDelete)
+            {
+                entity.Deleted = true;
+                _dbContext.Entry(entity).State = EntityState.Modified;
+                return await _dbContext.SaveChangesAsync();
+            }
+
             _dbContext.Set<T>().Remove(entity);
             return await _dbContext.SaveChangesAsync();
         }
 
         public override async Task<IReadOnlyList<T>> GetAllAsync(int page, int size)
         {
-            return await _dbContext.Set<T>().Skip((page - 1) * size).Take(size).AsNoTracking().ToListAsync();
+            return await _dbContext.Set<T>().Where(x => !x.Deleted)
+                .Skip((page - 1) * size).Take(size).AsNoTracking().ToListAsync();
         }
 
         public override async Task<IReadOnlyList<T>> GetAsync(Expression<Func<T, bool>> filter)
@@ -62,9 +72,21 @@ namespace Garcia.Persistence.EntityFramework
             return await _dbContext.SaveChangesAsync();
         }
 
-        public override async Task<long> DeleteManyAsync(Expression<Func<T, bool>> filter)
+        public override async Task<long> DeleteManyAsync(Expression<Func<T, bool>> filter, bool hardDelete = false)
         {
             var entities = _dbContext.Set<T>().Where(filter);
+
+            if(!hardDelete)
+            {
+                await entities.ForEachAsync(x =>
+                {
+                    x.Deleted = true;
+                    _dbContext.Entry(x).State = EntityState.Modified;
+                });
+
+                return await _dbContext.SaveChangesAsync();
+            }
+
             _dbContext.Set<T>().RemoveRange(entities);
             return await _dbContext.SaveChangesAsync();
         }
@@ -83,7 +105,7 @@ namespace Garcia.Persistence.EntityFramework
                     query = query.Include(property.Name);
             }
 
-            return await query.FirstOrDefaultAsync(x => x.Id == id);
+            return await query.FirstOrDefaultAsync(x => !x.Deleted && x.Id == id);
         }
     }
 }
